Fix linter and split up code more cleanly
