+++ packages/apolloschurchapp/src/ui/LikeButton/index.js
@@ -3,15 +3,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { get } from 'lodash';
 import { Query, Mutation } from 'react-apollo';
-import gql from 'graphql-tag';
 
 import Like from 'apolloschurchapp/src/ui/Like';
 import { track, events } from 'apolloschurchapp/src/analytics';
-import getAllLikedContent from 'apolloschurchapp/src/tabs/connect/getLikedContent';
-import { contentItemFragment } from 'apolloschurchapp/src/content-single/getContentItem';
 
 import updateLikeEntity from './updateLikeEntity';
 import getLikedContentItem from './getLikedContentItem';
+import updateLikedContent from './updateLikedContent';
 
 const GetLikeData = ({ itemId, children }) => (
   <Query query={getLikedContentItem} variables={{ itemId }}>
@@ -25,67 +23,6 @@ const GetLikeData = ({ itemId, children }) => (
 GetLikeData.propTypes = {
   itemId: PropTypes.string,
   children: PropTypes.func.isRequired,
-};
-
-const addItemToLikedContentList = ({ cache, item }) => {
-  try {
-    const data = cache.readQuery({ query: getAllLikedContent });
-    const fullItem = cache.readFragment({
-      id: `${item.__typename}:${item.id}`,
-      fragmentName: 'LikedContentItemParts',
-      fragment: gql`
-        fragment LikedContentItemParts on ContentItem {
-          ...contentItemFragment
-          __typename
-          id
-          coverImage {
-            name
-            sources {
-              uri
-            }
-          }
-          isLiked
-          parentChannel {
-            id
-            name
-            iconName
-          }
-          title
-        }
-        ${contentItemFragment}
-      `,
-    });
-    cache.writeQuery({
-      query: getAllLikedContent,
-      data: {
-        ...data,
-        getAllLikedContent: [fullItem, ...data.getAllLikedContent],
-      },
-    });
-  } catch (e) {
-    console.log(e);
-    // Most likely we haven't ran the `getAllLikedContent` query yet.
-    // We can safely exit.
-  }
-};
-
-const removeItemFromLikedContentList = ({ cache, item }) => {
-  try {
-    const data = cache.readQuery({ query: getAllLikedContent });
-    cache.writeQuery({
-      query: getAllLikedContent,
-      data: {
-        ...data,
-        getAllLikedContent: data.getAllLikedContent.filter(
-          (content) => content.id !== item.id
-        ),
-      },
-    });
-  } catch (e) {
-    console.log(e);
-    // Most likely we haven't ran the `getAllLikedContent` query yet.
-    // We can safely exit.
-  }
 };
 
 const UpdateLikeStatus = ({ itemId, item, isLiked, children }) => (
@@ -107,9 +44,7 @@ const UpdateLikeStatus = ({ itemId, item, isLiked, children }) => (
         },
       }
     ) => {
-      operation === 'Like'
-        ? addItemToLikedContentList({ cache, item })
-        : removeItemFromLikedContentList({ cache, item });
+      updateLikedContent({ liked: operation === 'Like', cache, item });
       cache.writeQuery({
         query: getLikedContentItem,
         data: {
