+++ packages/apolloschurchapp/src/ui/LikeButton/updateLikedContent.js
@@ -1,0 +1,74 @@
+import gql from 'graphql-tag';
+import getAllLikedContent from 'apolloschurchapp/src/tabs/connect/getLikedContent';
+import { contentItemFragment } from 'apolloschurchapp/src/content-single/getContentItem';
+
+const addItemToLikedContentList = ({ cache, item }) => {
+  try {
+    const data = cache.readQuery({ query: getAllLikedContent });
+    const fullItem = cache.readFragment({
+      id: `${item.__typename}:${item.id}`,
+      fragmentName: 'LikedContentItemParts',
+      fragment: gql`
+        fragment LikedContentItemParts on ContentItem {
+          ...contentItemFragment
+          __typename
+          id
+          coverImage {
+            name
+            sources {
+              uri
+            }
+          }
+          isLiked
+          parentChannel {
+            id
+            name
+            iconName
+          }
+          title
+        }
+        ${contentItemFragment}
+      `,
+    });
+    cache.writeQuery({
+      query: getAllLikedContent,
+      data: {
+        ...data,
+        getAllLikedContent: [fullItem, ...data.getAllLikedContent],
+      },
+    });
+  } catch (e) {
+    console.log(e);
+    // Most likely we haven't ran the `getAllLikedContent` query yet.
+    // We can safely exit.
+  }
+};
+
+const removeItemFromLikedContentList = ({ cache, item }) => {
+  try {
+    const data = cache.readQuery({ query: getAllLikedContent });
+    cache.writeQuery({
+      query: getAllLikedContent,
+      data: {
+        ...data,
+        getAllLikedContent: data.getAllLikedContent.filter(
+          (content) => content.id !== item.id
+        ),
+      },
+    });
+  } catch (e) {
+    console.log(e);
+    // Most likely we haven't ran the `getAllLikedContent` query yet.
+    // We can safely exit.
+  }
+};
+
+const updateLikedContent = ({ liked, cache, item }) => {
+  if (liked) {
+    addItemToLikedContentList({ cache, item });
+  } else {
+    removeItemFromLikedContentList({ cache, item });
+  }
+};
+
+export default updateLikedContent;
